==> Classification/1-train-check-classification.ipynb <==
import pandas as pd


df = pd.read_csv("released_dataset/train-classification.csv")

# Correct column names
df.columns = ["username", "label"]

# Lowercase label names.
df["label"] = df["label"].apply(str.lower)

# Remove any NA field.
print(f"Shape before `dropna`: {df.shape}")
df = df.dropna()
print(f"Shape after `dropna`: {df.shape}")

# Check there is no duplicate usernames.
len_unique_username = df["username"].unique().shape
print(f"Unique usernames amount: {len_unique_username}")

df.to_parquet("data/train-classification.parquet")

df = pd.read_parquet('data/train-classification.parquet')

df
