==> Classification/3-test-check-usernames.ipynb <==
import pandas as pd

df_train_classification = pd.read_parquet("data/train-classification.parquet")
trainable_usernames = df_train_classification["username"].tolist()

df_test = pd.read_csv("released_dataset/test-classification-round3.dat", header=None)
df_test.columns = ["username"]
test_usernames = df_test["username"].tolist()

print(f"Raw test usernames count: {len(test_usernames)}")


test_usernames_in_train = []
test_usernames_to_predict = []
# How many usernames in test set has its true labels in training set?
for test_username in test_usernames:
    if test_username in trainable_usernames:
        test_usernames_in_train.append(test_username)
    else:
        test_usernames_to_predict.append(test_username)

print(f"Usernames in test set that has correct labels in training data: {len(test_usernames_in_train)}")
print(f"Usernames in test set that has not correct labels in training data: {len(test_usernames_to_predict)}")



df_train = pd.read_parquet('data/training-dataset.parquet')

usernames_train_dataset = df_train['username'].tolist()

usernames_cant_predicted = []
for username in test_usernames_to_predict:
    if not username in usernames_train_dataset:
        usernames_cant_predicted.append(username)

print(f"There are {len(usernames_cant_predicted)} usernames that can't be predicted due to lack of data.")
print(f"These are {usernames_cant_predicted}")
df_train

# We can actually predict every username in the test set.

To do both the training and the test, we need embeddings of the usernames.

1. Usernames that are in the train set which has the correct category values. (To Train)
2. Useranames that are in the test set which do not have the correct category value. (To Test)

test_usernames_to_predict

train_usernames = df_train.dropna(subset=['label'])['username'].tolist()
train_usernames

all_usernames_to_embed = [*train_usernames, *test_usernames_to_predict]

df_username_biography = pd.read_parquet("data/username_biography.parquet")
df_username_caption = pd.read_parquet("data/username_caption.parquet")

df_uc = df_username_caption[df_username_caption["username"].isin(all_usernames_to_embed)]
df_ub = df_username_biography[df_username_biography["username"].isin(all_usernames_to_embed)]

df_uc.to_parquet('data/username_caption_2em.parquet')
df_ub.to_parquet('data/username_biography_2em.parquet')
