==> Classification/4-preprocess-text.ipynb <==
import pandas as pd
from tqdm.notebook import tqdm
tqdm.pandas()

# `Preprocessing` code is taken from ViralLab's GitHub repository:
# [TurkishBERTweet]
# Repository URL: https://github.com/ViralLab/TurkishBERTweet
from Preprocessor import preprocess

df_ub = pd.read_parquet('data/username_biography_2em.parquet')
df_ca = pd.read_parquet('data/username_caption_2em.parquet')

df_ub = df_ub.dropna()

df_ub['biography'] = df_ub['biography'].progress_apply(preprocess)
df_ub

df_ca['caption'] = df_ca['caption'].progress_apply(preprocess)
df_ca

df_ub.to_parquet('data/ub_2em_clean.parquet')
df_ca.to_parquet('data/uc_2em_clean.parquet')

df_ub = pd.read_parquet('data/ub_2em_clean.parquet')
df_ca = pd.read_parquet('data/uc_2em_clean.parquet')

#df_ub.to_csv('data/ub_2em_clean.csv')
#df_ca.to_csv('data/uc_2em_clean.csv')
