==> Classification/4.5-embedding-compute-colab.ipynb <==
!pip install torch torchvision torchaudio
!pip install peft
!pip install transformers
!pip install urlextract
!pip install pandas numpy

from tqdm.notebook import tqdm
tqdm.pandas()

import torch
from transformers import AutoTokenizer, AutoModel
device = torch.device("cuda" if torch.cuda.is_available() else "cpu")

print(device)

tokenizer = AutoTokenizer.from_pretrained("VRLLab/TurkishBERTweet")
turkishBERTweet = AutoModel.from_pretrained("VRLLab/TurkishBERTweet")

turkishBERTweet.to(device)

def create_embeddings(text):
    inputs = tokenizer(text, return_tensors="pt", truncation=True, padding="max_length", max_length=256).to(device)

    with torch.no_grad():
        outputs = turkishBERTweet(**inputs)

    last_hidden_states = outputs.last_hidden_state
    cls_embedding = last_hidden_states[:, 0, :]

    return cls_embedding.cpu().numpy().flatten().tolist()

import pandas

df_ub = pd.read_csv('ub_2em_clean.csv', index_col=0)
df_ub['embeds'] = df_ub['biography'].progress_apply(create_embeddings)
df_ub.to_csv('ub_em.csv')
df_ub

import pandas

df_uc = pd.read_csv('uc_2em_clean.csv', index_col=0)
df_uc["embeds"] = df_uc["caption"].progress_apply(create_embeddings)

df_uc.to_csv('uc_em.csv')
df_uc
