==> Classification/4.5-embedding-compute-openai-clean.ipynb <==
import pandas as pd




b0 = pd.read_json('openai-batch/response/batch_0.jsonl', lines=True)
b1 = pd.read_json('openai-batch/response/batch_1.jsonl', lines=True)
b2 = pd.read_json('openai-batch/response/batch_2.jsonl', lines=True)

from tqdm.notebook import tqdm
tqdm.pandas()


def extract_embedding(response):
    embed = response['body']['data'][0]['embedding']
    return embed

b0['embeds'] = b0['response'].progress_apply(extract_embedding)
b1['embeds'] = b1['response'].progress_apply(extract_embedding)
b2['embeds'] = b2['response'].progress_apply(extract_embedding)


FIELDS = ['custom_id','embeds']

df_all_em = pd.concat([b0[FIELDS],b1[FIELDS], b2[FIELDS]], axis=0)
df_all_em

df_all_em = df_all_em.reset_index(drop=True)[['embeds']]
df_all_em

df_ub = pd.read_parquet('data/username_biography_2em.parquet')
df_uc = pd.read_parquet('data/username_caption_2em.parquet')

df_ub = df_ub.reset_index(drop=True)
df_ub.columns = ['username','text']
df_ub = df_ub.dropna()

df_uc = df_uc.reset_index(drop=True)
df_uc.columns = ['username','text']
df_all_no_em = pd.concat([df_ub, df_uc], axis=0).reset_index(drop=True)


df_all_all = pd.concat([df_all_no_em, df_all_em], axis=1)
df_all_all

from consts import *

df_ub_em = df_all_all[:3253]
df_ub_em.columns = ['username','biography','embeds']
df_ub_em = df_ub_em.reset_index(drop=True)
df_ub_em.to_parquet(get_filename(TT.BIO, EMBED_PROVIDER.OPENAI))

df_uc_em = df_all_all[3253:]
df_uc_em.columns = ['username','caption','embeds']
df_uc_em = df_uc_em.reset_index(drop=True)
df_uc_em.to_parquet(get_filename(TT.CAP, EMBED_PROVIDER.OPENAI))
