==> Classification/4.5-embedding-compute-openai.ipynb <==
import pandas as pd
from tqdm.notebook import tqdm
tqdm.pandas()

df_ub = pd.read_parquet('data/username_biography_2em.parquet')

df_uc = pd.read_parquet('data/username_caption_2em.parquet')

df_ub = df_ub.reset_index(drop=True)
df_ub.columns = ['username','text']
df_ub = df_ub.dropna()


df_uc = df_uc.reset_index(drop=True)
df_uc.columns = ['username','text']


a = pd.concat([df_ub, df_uc], axis=0).reset_index(drop=True)
a

df_ub

a[:3474]
a


def create_batch_entry(row):

    custom_id  = row.name
    text = row['text']
    template = {"custom_id":str(custom_id), "method": "POST", "url": "/v1/embeddings", "body": {"model": "text-embedding-3-large", "input":text }}

    return template


a['be'] = a.progress_apply(create_batch_entry, axis=1)
a['index'] = a.progress_apply(lambda x: x.name, axis=1)
a

import pandas as pd
import numpy as np



chunk_size = 50000
chunks = np.array_split(a, 3)

for index, chunk in enumerate(chunks):
    chunk['be'].to_json(f'openai-batch/request/batch_{index}.json',lines=True, orient='records')

cc = chunks[0]
cc
