==> Classification/4.5-embedding-compute-st.ipynb <==
import pandas as pd
from sentence_transformers import SentenceTransformer
from tqdm.notebook import tqdm
tqdm.pandas()

model = SentenceTransformer("sentence-transformers/distiluse-base-multilingual-cased-v1", device='cuda')

df_username_bio = pd.read_parquet('data/username_biography_2em.parquet')
df_username_cap = pd.read_parquet('data/username_caption_2em.parquet')



df_username_bio['embeds'] = df_username_bio['biography'].progress_apply(lambda x: model.encode([x])[0] if x else None)

df_username_bio.to_parquet('data/ub_em_distil.parquet')

df_username_cap['embeds'] = df_username_cap['caption'].progress_apply(lambda x: model.encode([x])[0] if x else None)
df_username_cap.to_parquet('data/uc_em_distil.parquet')
