==> Classification/5-embedding-uc-combine-mean.ipynb <==
import pandas as pd
from tqdm.notebook import tqdm

tqdm.pandas()
from consts import *


F_IN = get_filename(TT.CAP, EMBED_PROVIDER.OPENAI)
F_OUT = get_filename(TT.CAP, EMBED_PROVIDER.OPENAI, MEAN_METHOD.MEAN)


df = pd.read_parquet(F_IN)
df

import numpy as np

usernames = df['username'].unique().tolist()

new_df = []

for username in tqdm(usernames):
    rows = df[df['username'] == username]
    embeds = []
    for index, row in rows.iterrows():
        embeds.append(row['embeds'])

    mean_embed = np.mean(embeds,axis=0)
    new_df.append((username, mean_embed))



new = pd.DataFrame(new_df)
new.columns = ['username','embeds']
new.to_parquet(F_OUT)
new
