==> Classification/5-embedding-uc-combine-w-centroid.ipynb <==
import pandas as pd
import numpy as np
from sklearn.metrics import pairwise_distances
from tqdm.notebook import tqdm
from consts import *
tqdm.pandas()


EMBED_FILE = get_filename(TT.CAP, EMBED_PROVIDER.OPENAI)
EMBED_FOUT = get_filename(TT.CAP, EMBED_PROVIDER.OPENAI, MEAN_METHOD.CENTROID)


df = pd.read_parquet(EMBED_FILE)
df

def combine_embeddings_with_relevance(embeddings):
    embeddings = np.array(embeddings)
    
    center_embedding = np.mean(embeddings, axis=0)
    
    def cosine_similarity(a, b):
        return np.dot(a, b) / (np.linalg.norm(a) * np.linalg.norm(b))
    
    similarities = np.array([cosine_similarity(embed, center_embedding) for embed in embeddings])
    
    weights = similarities - np.min(similarities)
    weights /= np.sum(weights)
    
    weighted_embedding = np.sum(embeddings.T * weights, axis=1)
    
    return weighted_embedding

usernames = df["username"].unique().tolist()

new_df = []

for username in tqdm(usernames):
    rows = df[df["username"] == username]
    embeds = []
    for index, row in rows.iterrows():
        embeds.append(row["embeds"])

    mean_embed = combine_embeddings_with_relevance(embeds)

    new_df.append((username, mean_embed))


new = pd.DataFrame(new_df)
new.columns = ['username','embeds']
new

new.to_parquet(EMBED_FOUT)
