==> Classification/6-embedding-combine-all.ipynb <==
import pandas as pd
from tqdm.notebook import tqdm
from consts import *

tqdm.pandas()


F_IN_BIO = get_filename(TT.BIO, EMBED_PROVIDER.TURKISH_BERT_TWEET)
F_IN_CAP = get_filename(TT.CAP, EMBED_PROVIDER.TURKISH_BERT_TWEET, MEAN_METHOD.MEAN)
F_OUT = get_filename(TT.ALL, EMBED_PROVIDER.TURKISH_BERT_TWEET, MEAN_METHOD.MEAN)

print(F_IN_BIO)
print(F_IN_CAP)
print(F_OUT)

BIO_RATIO = 0.1

df_ub = pd.read_parquet(F_IN_BIO)
df_uc = pd.read_parquet(F_IN_CAP)

df_uc

df_ub

df_combined = pd.merge(df_ub, df_uc, how='outer', on='username',suffixes=['_bio','_captions'])
df_combined = df_combined[['username','embeds_bio', 'embeds_captions']]
df_combined = df_combined.where(df_combined.notna(), None)


df_combined

import numpy as np


def embed_combinator(row):

    e_bio = row['embeds_bio']
    e_cap = row['embeds_captions']

    if e_bio is None and e_cap is None:
        return None


    if e_bio is None:
        return e_cap
    
    if e_cap is None:
        return e_bio

    return np.array(e_bio) * BIO_RATIO + np.array(e_cap) * (1-BIO_RATIO)



df_combined['embeds_wavg'] = df_combined.progress_apply(embed_combinator,axis=1)
df_combined = df_combined[['username','embeds_wavg']]
df_combined = df_combined.dropna()

df_combined['embeds_wavg'] = df_combined['embeds_wavg'].apply(lambda x: list(x))
df_combined

df_combined.to_parquet(F_OUT)
