==> Classification/8-test-xgbclassifier.ipynb <==
import pandas as pd
import xgboost as xgb


df_test = pd.read_csv('released_dataset/test-classification-round3.dat', header=None)
df_test.columns = ['username']
df_test

df_train = pd.read_parquet("data/training-dataset.parquet")
df_train

from sklearn.preprocessing import TargetEncoder

df_enum_lab = df_train[['label','category_enum']]
df_enum_lab = df_enum_lab.dropna(subset=["label"])

df_enum_lab["category_enum"] = df_enum_lab["category_enum"].fillna("Missing")

target_encoder = TargetEncoder()

target_encoder.fit(df_enum_lab["category_enum"].to_numpy().reshape(-1, 1), df_enum_lab["label"].tolist())

from consts import *

df_embeds = pd.read_parquet(get_filename(TT.ALL, EMBED_PROVIDER.OPENAI, MEAN_METHOD.CENTROID))
df_embeds

df_test_m1 = pd.merge(df_test,df_train, how='left', on='username')
df_test_m1

df_test_m2 = pd.merge(df_test_m1, df_embeds, how='left', on='username')
df_test_m2

df_test_correct = df_test_m2[~df_test_m2['label'].isna()]
df_test_correct = df_test_correct[['username','label']].reset_index(drop=True)
df_test_correct

df_test_to_predict = df_test_m2[~df_test_m2['username'].isin(df_test_correct['username'])].reset_index(drop=True)
df_test_to_predict = df_test_to_predict[['username','category_enum','embeds_wavg']]
df_test_to_predict

category_encoded_test = target_encoder.transform(df_test_to_predict["category_enum"].to_numpy().reshape(-1, 1))
df_test_to_predict["category_encoded"] = [row.tolist() for row in category_encoded_test]
df_test_to_predict



X_test = pd.concat(
    [
        pd.DataFrame(df_test_to_predict["category_encoded"].tolist()),
        pd.DataFrame(df_test_to_predict["embeds_wavg"].tolist()),
    ],
    axis=1,
)

X_test

X_test_numpy = X_test.to_numpy()

import xgboost as xgb

dtest = xgb.DMatrix(X_test_numpy)

model = xgb.Booster()
model.load_model('trained_models/xgb_openai_centroid.json')

y_pred = model.predict(dtest).astype(int)
y_pred

from sklearn.preprocessing import LabelEncoder

labels = [
    "entertainment",
    "food",
    "travel",
    "health and lifestyle",
    "mom and children",
    "fashion",
    "tech",
    "sports",
    "art",
    "gaming",
]

label_encoder = LabelEncoder()
label_encoder.fit(labels)

y_pred_labels = label_encoder.inverse_transform(y_pred)
y_pred_labels

df_test_predicted = pd.concat([df_test_to_predict['username'], pd.DataFrame(y_pred_labels)], axis=1)
df_test_predicted.columns = ['username','label']
df_test_predicted

df_predicted = pd.concat([df_test_predicted, df_test_correct], axis=0).reset_index(drop=True)
df_predicted

label_conv = {
    "entertainment":'Entertainment',
    "food":'Food',
    "travel":'Travel',
    "health and lifestyle":'Health and Lifestyle',
    "mom and children":'Mom and Children',
    "fashion":'Fashion',
    "tech":'Tech',
    "sports":'Sports',
    "art":'Art',
    "gaming":'Gaming',
}

def correct_label(label):

    return label_conv[label]

df_predicted['label_correct'] = df_predicted['label'].apply(correct_label)
df_predicted = df_predicted.drop(columns=['label'])
df_predicted.columns = ['username','label']
df_predicted


import json

result_dict = df_predicted.set_index('username')['label'].to_dict()
open('predictions/prediction-classification-round3.json','w').write(json.dumps(result_dict,indent=4))
